requirements.txt
annotated-types==0.7.0
anthropic==0.46.0
anyio==4.9.0
beautifulsoup4==4.13.3
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
distro==1.9.0
fastapi==0.115.11
filelock==3.13.1
filetype==1.2.0
fsspec==2024.6.1
ftfy==6.3.1
google-auth==2.38.0
google-genai==1.5.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
huggingface-hub==0.29.3
idna==3.10
Jinja2==3.1.4
jiter==0.9.0
joblib==1.4.2
markdown2==2.5.3
markdownify==0.13.1
marker-pdf==1.6.1
MarkupSafe==2.1.5
mpmath==1.3.0
networkx==3.3
numpy==2.1.2
opencv-python-headless==4.11.0.86
packaging==24.2
pdftext==0.6.2
pillow==10.4.0
platformdirs==4.3.6
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.11.0b2
pydantic-settings==2.8.1
pydantic_core==2.32.0
pypdfium2==4.30.0
python-dotenv==1.0.1
PyYAML==6.0.2
RapidFuzz==3.12.2
regex==2024.11.6
requests==2.32.3
rsa==4.9
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.2
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
starlette==0.46.1
surya-ocr==0.13.0
sympy==1.13.1
threadpoolctl==3.6.0
tokenizers==0.21.1
torch==2.6.0+cu124
torchaudio==2.6.0+cu124
torchvision==0.21.0+cu124
tqdm==4.67.1
transformers==4.49.0
typing-inspection==0.4.0
typing_extensions==4.12.2
urllib3==2.3.0
uvicorn==0.34.0
wcwidth==0.2.13
websockets==14.2
